requirements.txt
black==22.3.0
catboost==1.0.6
category-encoders==2.4.1
flake8==4.0.1
hydra-core==1.2.0
ipykernel==6.13.0
isort==5.10.1
lightgbm==3.3.2
matplotlib==3.5.2
numpy==1.22.4
optuna==2.10.0
pandas==1.4.2
plotly==5.8.0
scikit-learn==1.1.1
seaborn==0.11.2
tqdm==4.64.0
wandb==0.12.17
xgboost==1.6.1